requirements.txt
cohere==4.11.2
sentence-transformers==2.7.0
requests==2.31.0
streamlit==1.42.0
langchain==0.2.1
pinecone-client==4.1.0
accelerate==0.30.1
transformers==4.41.1
langchain_community==0.2.1
langdetect==1.0.9
beautifulsoup4==4.12.3
chromadb==0.5.3
docx2txt==0.8
python-docx==0.8.11

#PyPDF2==3.0.1
#pypdf==4.2.0
#arabic-reshaper==3.0.0
#python-bidi==0.6.6
#python-docx==1.1.2